adjusted structure by moving __init__ into Backend, added vercel.json, ran npm run build, altere requirements.txt
