+++ Backend/app.py
@@ -1,0 +1,9 @@
+from __init__ import create_app
+from config import DevelopmentConfig
+from flask_cors import CORS
+
+app = create_app(config_class=DevelopmentConfig)
+CORS(app)
+
+if __name__ == '__main__':
+    app.run(host='0.0.0.0', port=5000, debug=True)
