+++ Backend/tests/test_apply.py
@@ -9,7 +9,7 @@ from flask_testing import TestCase
 
 from src.models.employees import Employees
 from src.models.wfh_requests import WFH_Requests
-from src.__init__ import db, create_app
+from __init__ import db, create_app
 from config import TestingConfig
 
 class BaseTestCase(TestCase):
